feat: create bn_add_cli crate and move CLI to separate package
