+++ Cargo.toml
@@ -3,11 +3,15 @@ name = "bn_add"
 version = "0.1.0"
 edition = "2021"
 
+[workspace]
+members = [
+    "bn_add_cli"
+]
+
 [dependencies]
 vstd = { git = "https://github.com/verus-lang/verus", rev = "9b557d70faab565829c0642c4d56609a44814d82" }
 builtin = { git = "https://github.com/verus-lang/verus", rev = "9b557d70faab565829c0642c4d56609a44814d82" }
 builtin_macros = { git = "https://github.com/verus-lang/verus", rev = "9b557d70faab565829c0642c4d56609a44814d82" }
-clap = { version = "4.5", features = ["derive"] }
 
 [package.metadata.verus]
 verify = true
