+++ bn_add_cli/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "bn_add_cli"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+bn_add = { path = "../" }
+clap = { version = "4.5", features = ["derive"] }
